fix!: Separate sentryiris into its own module.
