+++ sentryiris/go.mod
@@ -1,0 +1,8 @@
+module github.com/getsentry/sentry-go/sentryiris
+
+go 1.14
+
+require (
+	github.com/getsentry/sentry-go v0.11.0
+	github.com/kataras/iris/v12 v12.1.8
+)
